third_party/WebKit/LayoutTests/typedcssom/usecounters/declared-stylePropertyMap.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/testhelper.js"></script>
<body>
<script>
'use strict';

test(t => {
  const CSSTypedOMStylePropertyMap = 2334; // From UseCounter.h

  assert_false(internals.isUseCounted(document, CSSTypedOMStylePropertyMap));

  let styleMap = createDeclaredStyleMap(t, '');

  assert_true(internals.isUseCounted(document, CSSTypedOMStylePropertyMap));
}, 'We UseCount when we obtain a declared style property map');

</script>
